Zephyr Painters
Idea Events. Currently, my wife and I are incorporated as "Zephyr Painters, LLC." Along with many of our friends in the area, including Ellie & Miles Wilson, we are interested in proposing an "Northeastern Tennessee Alternative Triathlon." On 8th February 2022, our city council has a mandated meeting as instructed by the county to proceed with the project. 

A new annual event for this area it includes a two mile swim across Whatuaga Lake (Reservoir) ending at the docks at the "Captain's Table," a mountain bike challenge of approximately seventy-five miles along the scenic Appalachian Trail, and a final test of endurance including a half-marathon ending either in Bristol, TN on Main Street. A music festival in both Bristol's will follow the Triathlon.